pyproject.toml
[project]
name = "oppelganger"
version = "0.0.0"
description = "Цепь из LLM, TTS, LipSync, GfpGan образующие движок личности"
authors = [
    { name = "func", email = "rtem903134@gmail.com" },
	{ name = "Kamillaova", email = "me@kamillaova.dev" },
]
dependencies = [
    "llama-cpp-python>=0.2.26",
    "TTS>=0.22.0",
    "torchaudio>=2.1.2",
    "opencv-python>=4.9.0.80",
    "mediapipe>=0.10.9",
    "torch>=2.1.2",
    "gfpgan>=1.3.8",
    "runpod>=1.5.0",
    "pydantic>=2.5.3",
    "boto3>=1.34.17",
    "boto3-stubs[s3]>=1.34.18",
    "fastapi>=0.110.3",
    "pycuda>=2024.1",
    "mypy-boto3-s3>=1.34.91",
    "uvicorn>=0.29.0",
]
requires-python = ">=3.11,<3.12"
readme = "README.md"
license = "LicenseRef-Oppelganger"

[tool.pdm.dev-dependencies]
lint = [
	"mypy>=1.8.0",
]

[tool.pdm]
package-type = "application"
